backend/test/test_config.py
import os
import unittest
from flask import current_app
from flask_testing import TestCase
from manage import app
from backend.main.config import basedir


class TestDevelopmentConfig(TestCase):
    def create_app(self):
        app.config.from_object('backend.main.config.DevelopmentConfig')
        return app

    def test_app_is_development(self):
        self.assertTrue(app.config['SECRET_KEY'] is 'lcgFGnompC')
        self.assertTrue(app.config['DEBUG'] is True)
        self.assertFalse(current_app is None)
        self.assertTrue(
            app.config['SQLALCHEMY_DATABASE_URI'] == 'sqlite:///' + os.path.join(basedir, 'rss_reader.db')
        )


class TestTestingConfig(TestCase):
    def create_app(self):
        app.config.from_object('backend.main.config.TestingConfig')
        return app

    def test_app_is_testing(self):
        self.assertTrue(app.config['SECRET_KEY'] is 'lcgFGnompC')
        self.assertTrue(app.config['DEBUG'])
        self.assertTrue(
            app.config['SQLALCHEMY_DATABASE_URI'] == 'sqlite:///' + os.path.join(basedir, 'rss_reader_test.db')
        )


if __name__ == '__main__':
    unittest.main()
